+++ lib/camp/providers/comment_model.dart
@@ -4,14 +4,14 @@ import 'package:koye_kos/models/comment.dart';
 
 abstract class RatingProvider {
   void onRated(double score);
-  double get score;
+  int get score;
 }
 
 class CommentModel extends RatingProvider with ChangeNotifier {
   final String originalText;
-  final double originalScore;
+  final int originalScore;
   String commentText = '';
-  double campScore = 0;
+  int campScore = 0;
   CommentModel({this.originalText, this.originalScore}) {
     commentText = originalText ?? '';
     campScore = originalScore ?? 0;
@@ -36,11 +36,11 @@ class CommentModel extends RatingProvider with ChangeNotifier {
   }
 
   @override
-  double get score => campScore ?? 0;
+  int get score => campScore ?? 0;
 
   @override
   void onRated(double score) {
-    campScore = score;
+    campScore = score.toInt();
     notifyListeners();
   }
 }
